Refactor import paths for components in Reservations.jsx and Reserve.jsx
